requirements.txt
altgraph==0.17
clipboard==0.0.4
importlib-metadata==3.10.0
macholib==1.14
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pygame==2.0.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyperclip==1.8.2
tqdm==4.59.0
typing-extensions==3.7.4.3
zipp==3.4.1
